Merge latest develop

+++ conans/client/packager.py
@@ -1,16 +1,18 @@
-from conans.util.files import mkdir, save, rmdir
 import os
+import shutil
+
+from conans.client import tools
+from conans.util.files import mkdir, save, rmdir
 from conans.util.log import logger
 from conans.paths import CONANINFO, CONAN_MANIFEST
-from conans.errors import ConanException, format_conanfile_exception
-from conans.model.build_info import DEFAULT_RES, DEFAULT_BIN, DEFAULT_LIB, DEFAULT_INCLUDE
-import shutil
-from conans.client.file_copier import FileCopier
+from conans.errors import ConanException, ConanExceptionInUserConanfileMethod, conanfile_exception_formatter
 from conans.model.manifest import FileTreeManifest
 from conans.client.output import ScopedOutput
+from conans.client.file_copier import FileCopier
 
 
-def create_package(conanfile, build_folder, package_folder, output):
+def create_package(conanfile, source_folder, build_folder, package_folder, install_folder,
+                   output, local=False, copy_info=False):
     """ copies built artifacts, libs, headers, data, etc from build_folder to
     package folder
     """
@@ -19,54 +21,64 @@ def create_package(conanfile, build_folder, package_folder, output):
     # Make the copy of all the patterns
     output.info("Generating the package")
     output.info("Package folder %s" % (package_folder))
-    conanfile.copy = FileCopier(build_folder, package_folder)
 
-    def wrap(dst_folder):
-        def new_method(pattern, src=""):
-            conanfile.copy(pattern, dst_folder, src)
-        return new_method
-
-    conanfile.copy_headers = wrap(DEFAULT_INCLUDE)
-    conanfile.copy_libs = wrap(DEFAULT_LIB)
-    conanfile.copy_bins = wrap(DEFAULT_BIN)
-    conanfile.copy_res = wrap(DEFAULT_RES)
     try:
-        conanfile.package_folder = package_folder
-        conanfile.package()
         package_output = ScopedOutput("%s package()" % output.scope, output)
-        conanfile.copy.report(package_output, warn=True)
+        output.highlight("Calling package()")
+        conanfile.package_folder = package_folder
+        conanfile.source_folder = source_folder
+        conanfile.install_folder = install_folder
+        conanfile.build_folder = build_folder
+
+        def recipe_has(conanfile, attribute):
+            return attribute in conanfile.__class__.__dict__
+
+        if source_folder != build_folder:
+            conanfile.copy = FileCopier(source_folder, package_folder, build_folder)
+            with conanfile_exception_formatter(str(conanfile), "package"):
+                with tools.chdir(source_folder):
+                    conanfile.package()
+            warn = recipe_has(conanfile, "package")
+            conanfile.copy.report(package_output, warn=warn)
+
+        conanfile.copy = FileCopier(build_folder, package_folder)
+        with tools.chdir(build_folder):
+            with conanfile_exception_formatter(str(conanfile), "package"):
+                conanfile.package()
+        warn = recipe_has(conanfile, "build") and recipe_has(conanfile, "package")
+        conanfile.copy.report(package_output, warn=warn)
     except Exception as e:
-        os.chdir(build_folder)
-        try:
-            rmdir(package_folder)
-        except Exception as e_rm:
-            output.error("Unable to remove package folder %s\n%s" % (package_folder, str(e_rm)))
-            output.warn("**** Please delete it manually ****")
+        if not local:
+            os.chdir(build_folder)
+            try:
+                rmdir(package_folder)
+            except Exception as e_rm:
+                output.error("Unable to remove package folder %s\n%s" % (package_folder, str(e_rm)))
+                output.warn("**** Please delete it manually ****")
 
-        msg = format_conanfile_exception(output.scope, "package", e)
-        raise ConanException(msg)
+        if isinstance(e, ConanExceptionInUserConanfileMethod):
+            raise
+        raise ConanException(e)
 
-    _create_aux_files(build_folder, package_folder)
+    _create_aux_files(install_folder, package_folder, conanfile, copy_info)
     output.success("Package '%s' created" % os.path.basename(package_folder))
 
 
-def generate_manifest(package_folder):
+def _create_aux_files(install_folder, package_folder, conanfile, copy_info):
+    """ auxiliary method that creates CONANINFO and manifest in
+    the package_folder
+    """
+    logger.debug("Creating config files to %s" % package_folder)
+    if copy_info:
+        try:
+            shutil.copy(os.path.join(install_folder, CONANINFO), package_folder)
+        except IOError:
+            raise ConanException("%s does not exist inside of your %s folder. "
+                                 "Try to re-build it again to solve it."
+                                 % (CONANINFO, install_folder))
+    else:
+        save(os.path.join(package_folder, CONANINFO), conanfile.info.dumps())
+
     # Create the digest for the package
     digest = FileTreeManifest.create(package_folder)
     save(os.path.join(package_folder, CONAN_MANIFEST), str(digest))
-
-
-def _create_aux_files(build_folder, package_folder):
-    """ auxiliary method that creates CONANINFO in
-    the package_folder
-    """
-    try:
-        logger.debug("Creating config files to %s" % package_folder)
-        shutil.copy(os.path.join(build_folder, CONANINFO), package_folder)
-
-        # Create the digest for the package
-        generate_manifest(package_folder)
-
-    except IOError:
-        raise ConanException("%s does not exist inside of your %s folder. Try to re-build it again"
-                             " to solve it." % (CONANINFO, build_folder))
